+++ src/main/java/com/zen/fogman/common/entity/ModEntities.java
@@ -3,6 +3,7 @@ package com.zen.fogman.common.entity;
 import com.zen.fogman.common.ManFromTheFog;
 import com.zen.fogman.common.entity.the_man.TheManEntity;
 import com.zen.fogman.common.entity.the_man.TheManEntityHallucination;
+import com.zen.fogman.common.entity.the_man.TheManSpitEntity;
 import net.fabricmc.fabric.api.object.builder.v1.entity.FabricDefaultAttributeRegistry;
 import net.fabricmc.fabric.api.object.builder.v1.entity.FabricEntityTypeBuilder;
 import net.minecraft.entity.EntityDimensions;
@@ -23,6 +24,14 @@ public class ModEntities {
             Registries.ENTITY_TYPE,
             new Identifier(ManFromTheFog.MOD_ID, "the_man_hallucination"),
             FabricEntityTypeBuilder.create(SpawnGroup.MONSTER, TheManEntityHallucination::new).dimensions(TheManEntity.HITBOX_SIZE).build()
+    );
+
+    // Projectiles
+    public static final EntityType<TheManSpitEntity> THE_MAN_SPIT = Registry.register(
+            Registries.ENTITY_TYPE,
+            new Identifier(ManFromTheFog.MOD_ID,"the_man_spit"),
+            FabricEntityTypeBuilder.<TheManSpitEntity>create(SpawnGroup.MISC, TheManSpitEntity::new).dimensions(EntityDimensions.fixed(0.5f,0.5f))
+                    .build()
     );
 
     public static void register() {