+++ src/main/java/com/zen/fogman/common/entity/the_man/states/ChaseState.java
@@ -3,6 +3,7 @@ package com.zen.fogman.common.entity.the_man.states;
 import com.zen.fogman.common.entity.the_man.TheManEntity;
 import com.zen.fogman.common.entity.the_man.TheManPredicates;
 import com.zen.fogman.common.entity.the_man.TheManStatusEffects;
+import com.zen.fogman.common.gamerules.ModGamerules;
 import com.zen.fogman.common.other.Util;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.server.network.ServerPlayerEntity;
@@ -13,10 +14,14 @@ public class ChaseState extends AbstractState {
     public static final double LUNGE_COOLDOWN = 30;
     public static final double LUNGE_CHANCE = 0.4;
 
+    public static final double SPIT_COOLDOWN = 15;
+    public static final double SPIT_CHANCE = 0.45;
+
     public static final double HALLUCINATION_COOLDOWN = 60;
     public static final double HALLUCINATION_CHANCE = 0.1;
 
     private long lungeCooldown = Util.secToTick(LUNGE_COOLDOWN);
+    private long spitCooldown = Util.secToTick(SPIT_COOLDOWN);
     private long hallucinationCooldown = Util.secToTick(HALLUCINATION_COOLDOWN);
 
     public ChaseState(TheManEntity mob) {
@@ -33,9 +38,13 @@ public class ChaseState extends AbstractState {
 
         this.mob.breakBlocksAround();
 
-        this.mob.addEffectToClosePlayers(serverWorld, TheManStatusEffects.DARKNESS);
-        this.mob.addEffectToClosePlayers(serverWorld, TheManStatusEffects.SPEED);
-        this.mob.addEffectToClosePlayers(serverWorld, TheManStatusEffects.NIGHT_VISION);
+        if (serverWorld.getGameRules().getBoolean(ModGamerules.MAN_GIVE_DARKNESS_EFFECT)) {
+            this.mob.addEffectToClosePlayers(serverWorld, TheManStatusEffects.DARKNESS);
+            this.mob.addEffectToClosePlayers(serverWorld, TheManStatusEffects.NIGHT_VISION);
+        }
+        if (serverWorld.getGameRules().getBoolean(ModGamerules.MAN_GIVE_SPEED_EFFECT)) {
+            this.mob.addEffectToClosePlayers(serverWorld, TheManStatusEffects.SPEED);
+        }
 
         this.mob.getLookControl().lookAt(target,30f,30f);
         this.mob.moveTo(target,1.0);
@@ -45,6 +54,13 @@ public class ChaseState extends AbstractState {
             if (Math.random() < LUNGE_CHANCE) {
                 this.mob.setLunging(false);
                 this.mob.lunge(target,0.6);
+            }
+        }
+
+        if (this.mob.distanceTo(target) > 15 && --this.spitCooldown <= 0L) {
+            this.spitCooldown = Util.secToTick(SPIT_COOLDOWN);
+            if (Math.random() < SPIT_CHANCE) {
+                this.mob.spitAt(target);
             }
         }