+++ src/main/java/com/zen/fogman/common/gamerules/ModGamerules.java
@@ -14,6 +14,9 @@ public class ModGamerules {
 
     public static final GameRules.Key<GameRules.BooleanRule> MAN_CAN_SPAWN_IN_DAY = createRule("manCanSpawnInDay", GameRules.Category.SPAWNING, false);
 
+    public static final GameRules.Key<GameRules.BooleanRule> MAN_GIVE_DARKNESS_EFFECT = createRule("manShouldGiveDarknessEffect", GameRules.Category.MOBS, true);
+    public static final GameRules.Key<GameRules.BooleanRule> MAN_GIVE_SPEED_EFFECT = createRule("manShouldGiveSpeedEffect", GameRules.Category.MOBS, true);
+
     public static final GameRules.Key<GameRules.IntRule> MAN_MIN_SPAWN_RANGE = createRule("manMinSpawnRange", GameRules.Category.SPAWNING,20,20,120);
     public static final GameRules.Key<GameRules.IntRule> MAN_MAX_SPAWN_RANGE = createRule("manMaxSpawnRange", GameRules.Category.SPAWNING,60,20,120);